Add Tauri & Leptos template (run with `overmind start`)
I used code from this template:

https://gitlab.com/cristofa/tauri-leptos-template

But I had to workaround a bug in cargo-tauri, trunk, and/or rust-analyzer (not
sure) that kept rebuilding the project from scratch, every time. So I created
two workspaces, one for the backand (which runs native code) and another for the
frontend (which runs wasm), to separate their target dirs; and further separated
them by having rust-analyzer build into a subdirectory target/rust-analyzer.
That way they don't step on each other's toes. (See the urls in a comment on the
file rust-analyzer.toml for some more details)